Restructured interpreter module and its dependencies. Cleaned up test code.
