+++ infrastructure/app/Interpreter/NullInterpreter.php
@@ -1,4 +1,4 @@
-<?php namespace Infrastructure\App\Interpreter\InterpreterPattern;
+<?php namespace Infrastructure\App\Interpreter;
 
 use App\Interpreter\Context;
