+++ tests/Interpreter/InterpreterTest.php
@@ -1,21 +1,21 @@
 <?php
 
-require_once "EventStore.php";
+require_once "InvariantRepository/Fail.php";
+require_once "InvariantRepository/Pass.php";
 
-use App\Interpreter\Interpreter;
-use App\Interpreter\InvariantException;
+use App\Interpreter\InvariantRepository;
 use App\Interpreter\Context;
+use App\Interpreter\InvariantException;
+use Infrastructure\App\Interpreter\CommandHandler;
+use Infrastructure\App\Interpreter\Handler;
 
-//PrototypeInterpreter
-abstract class AbstractTest extends TestCase
-{   
+class InterpreterTest extends TestCase
+{
     protected $ast;
-    protected $event_store;
     
     public function setUp()
     {
         $this->ast = $this->ast();
-        $this->event_store = $this->event_store();
         
         $event = new \stdClass();
         $event->id = 'event_id';
@@ -24,19 +24,9 @@ abstract class AbstractTest extends TestCase
     
     private function ast()
     {
-        $ast_file = base_path('tests/PrototypeInterpreter/ast.json');
+        $ast_file = base_path('tests/Interpreter/ast.json');
         return json_decode(file_get_contents($ast_file));
     }
-    
-    private function event_store()
-    {
-        return new EventStore();
-    }
-    
-    /**
-     * @return Interpreter
-     */
-    abstract protected function build_fires_events_interpreter();
     
     private function command()
     {
@@ -52,29 +42,38 @@ abstract class AbstractTest extends TestCase
         $context = $context->set_property('command', $this->command());
         return $context;
     }
-   
+    
+    /**
+     * @return Interpreter
+     */
+    protected function build_fires_events_interpreter()
+    {
+        $this->app()->bind(InvariantRepository::class, Pass::class);
+        $handler_factory = $this->app()->make(Handler\Factory::class);
+        return new CommandHandler($handler_factory, $this->ast->handler);
+    }
+    
     public function test_interpreter_fires_events()
     {
         $interpreter = $this->build_fires_events_interpreter();
-        
-        $events = $interpreter->interpret($this->context());
-        
+        $events = $interpreter->interpret($this->context());    
         $this->assertEquals($this->expected_events, $events);  
     }
     
     /**
      * @return Interpreter
      */
-    abstract protected function build_fails_on_invariants_interpreter();
+    protected function build_fails_on_invariants_interpreter()
+    {
+        $this->app()->bind(InvariantRepository::class, Fail::class);
+        $handler_factory = $this->app()->make(Handler\Factory::class);
+        return new CommandHandler($handler_factory, $this->ast->handler);
+    }
     
     public function test_interpreter_fails_on_invariants()
     {
         $this->setExpectedException(InvariantException::class);
-        
         $interpreter = $this->build_fails_on_invariants_interpreter();
-        
         $interpreter->interpret($this->context());
-    }
+    }    
 }
-
-
