+++ tests/Interpreter/InvariantRepository/Fail.php
@@ -4,6 +4,6 @@ class Fail implements \App\Interpreter\InvariantRepository
 {
     public function fetch($id)
     {
-        return new \Infrastructure\App\Interpreter\InterpreterPattern\Invariant(true);
+        return new \Infrastructure\App\Interpreter\Invariant(true);
     }
 }
