+++ tests/Interpreter/InvariantRepository/Pass.php
@@ -4,6 +4,6 @@ class Pass implements \App\Interpreter\InvariantRepository
 {
     public function fetch($id)
     {
-        return new \Infrastructure\App\Interpreter\InterpreterPattern\Invariant(false);
+        return new \Infrastructure\App\Interpreter\Invariant(false);
     }
 }
